cpukit/score/cpu/microblaze/cpu.c: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ cpukit/score/cpu/microblaze/cpu.c
@@ -138,7 +138,7 @@ void _CPU_Exception_frame_print( const CPU_Exception_frame *ctx )
           ( ctx->msr & MICROBLAZE_MSR_IE ) ? "IE " : "" );
 
   const char *esr_ec_txt = "?";
-  int         exception_ind = 0;
+  size_t      exception_ind = 0;
   int         esr_ec = ctx->esr & 0x1f;
   for ( ; exception_ind < sizeof( esr_ec_codes ) / sizeof( esr_ec_codes[ 0 ] );
         exception_ind++ ) {
